[llvm-profdata][SampleProfileLoader] Pointer to GUIDToFuncNameMap should not be copied in every FunctionSample
When compiling each module, the GUIDToFuncNameMap is created to lookup
original function name (in the module) with its MD5 value. A pointer to
this map is being stored in every FunctionSample, which is not necessary
because they are all the same within the same module.
Such lookup should be fully handled within SampleProfileLoader class,
which also make SampleProfiles (mapping MD5 to FunctionSamples) slightly
more memory and CPU efficient by eliminating a class member.