[NFC] [SROA] Cache Basic Block Predecessors and Successors while Performing Mem2Reg
The CFG is repeatedly (partially) traversed when analyzing which BB does
an Alloca live. Profiling show that the current method getting
predecessors and successors of a BB (which returns a iterator range) is
a bottleneck in SROA pass because iterator increment is not trivial.
The result should be materialized and cached since many Allocas have
overlapping living BBs.